# Conflicts:
#	WooCommerce/WooCommerce.xcodeproj/project.pbxproj

+++ Networking/Networking/Mapper/ProductReviewListMapper.swift
@@ -7,7 +7,7 @@ struct ProductReviewListMapper: Mapper {
     ///
     /// We're injecting this field via `JSONDecoder.userInfo` because SiteID is not returned in any of the Product Endpoints.
     ///
-    let siteID: Int
+    let siteID: Int64
 
     /// (Attempts) to convert a dictionary into [Product].
     ///
