+++ Networking/Networking/Remote/OrderStatsRemoteV4.swift
@@ -15,7 +15,7 @@ public final class OrderStatsRemoteV4: Remote {
     ///
     /// Note: by limiting the return values with the `_fields` param, we shrink the response size by over 90%! (~40kb to ~3kb)
     ///
-    public func loadOrderStats(for siteID: Int,
+    public func loadOrderStats(for siteID: Int64,
                                unit: StatsGranularityV4,
                                earliestDateToInclude: String,
                                latestDateToInclude: String,
